+++ packages/create/src/formatting.ts
@@ -18,6 +18,7 @@ export const dim = formatter(2, 22);
 export const red = formatter(31, 39);
 export const green = formatter(32, 39);
 export const blue = formatter(34, 39);
+export const yellow = formatter(33, 39);
 
 export function visibleWidthOf(str: string) {
     let i = 0,
@@ -45,11 +46,30 @@ export function visibleWidthOf(str: string) {
     return length;
 }
 
-export function fittedTextOf(text: string, width: number) {
+export function fittedTextOf(text: string, width = stdout.columns) {
     if (width <= 20) {
         return text;
     }
 
+    const lines = text.split("\n");
+    const paragraphs = Array<string>();
+    for (let line of lines) {
+        const paddedLineMatch = line.match(/^(\s*)(.*)/);
+        let padding;
+        if (paddedLineMatch) {
+            padding = paddedLineMatch[1];
+            line = paddedLineMatch[2];
+        } else {
+            padding = "";
+        }
+
+        paragraphs.push(padding + wrap(line, width));
+    }
+
+    return paragraphs.join("\n");
+}
+
+function wrap(text: string, width = stdout.columns) {
     const words = text.trim().split(/\s+/);
     let lineWidth = 0;
     let line = "";