+++ packages/nodejs/src/net/register.ts
@@ -4,17 +4,28 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-import { Network, NoProviderError, singleton } from "#general";
+import { Boot, Network, NoProviderError, singleton } from "#general";
 import { NodeJsNetwork } from "./NodeJsNetwork.js";
 
-// Check if Network singleton is already registered and auto register if not
-try {
-    Network.get();
-} catch (error) {
-    NoProviderError.accept(error);
+let closeNetwork: undefined | (() => void);
 
-    Network.get = singleton(() => new NodeJsNetwork());
+Boot.init(() => {
+    closeNetwork?.();
 
-    // Ensure network gets cleaned up on exit
-    process.on("beforeExit", () => void Network.get().close());
-}
+    // Check if Network singleton is already registered and auto register if not
+    try {
+        Network.get();
+    } catch (error) {
+        NoProviderError.accept(error);
+
+        Network.get = singleton(() => new NodeJsNetwork());
+
+        closeNetwork = () => {
+            void Network.get().close();
+            process.off("beforeExit", close);
+        };
+
+        // Ensure network gets cleaned up on exit
+        process.on("beforeExit", closeNetwork);
+    }
+});