+++ packages/react-native/src/environment/register.ts
@@ -4,7 +4,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-import { Environment } from "@matter/general";
+import { Boot, Environment } from "@matter/general";
 import { ReactNativeEnvironment } from "./ReactNativeEnvironment.js";
 
 // Mainly import TextEncoder and TextDecoder polyfill
@@ -13,4 +13,6 @@ import { polyfill } from "react-native-polyfill-globals/src/encoding";
 
 polyfill();
 
-Environment.default = ReactNativeEnvironment();
+Boot.init(() => {
+    Environment.default = ReactNativeEnvironment();
+});