+++ packages/react-native/src/net/register.ts
@@ -4,14 +4,16 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-import { Network, NoProviderError, singleton } from "@matter/general";
+import { Boot, Network, NoProviderError, singleton } from "@matter/general";
 import { NetworkReactNative } from "./NetworkReactNative.js";
 
-// Check if Network singleton is already registered and auto register if not
-try {
-    Network.get();
-} catch (error) {
-    NoProviderError.accept(error);
-    Network.get = singleton(() => new NetworkReactNative());
-    // use net export closeNetwork() to close the network
-}
+Boot.init(() => {
+    // Check if Network singleton is already registered and auto register if not
+    try {
+        Network.get();
+    } catch (error) {
+        NoProviderError.accept(error);
+        Network.get = singleton(() => new NetworkReactNative());
+        // use net export closeNetwork() to close the network
+    }
+});