+++ packages/testing/src/global-definitions.ts
@@ -10,6 +10,7 @@
 import Chai from "chai";
 import ChaiAsPromised from "chai-as-promised";
 import { browserSetup, generalSetup } from "./mocha.js";
+import { bootSetup } from "./mocks/boot.js";
 import { cryptoSetup } from "./mocks/crypto.js";
 import { TheMockLogger, loggerSetup } from "./mocks/logging.js";
 import { timeSetup } from "./mocks/time.js";
@@ -21,6 +22,7 @@ Object.assign(globalThis, {
     expect: Chai.expect,
 
     MatterHooks: {
+        bootSetup,
         loggerSetup,
         timeSetup,
         cryptoSetup,