+++ packages/testing/src/mocks/index.ts
@@ -6,5 +6,6 @@
 
 import "./environment.js";
 import "./logging.js";
+export * from "./boot.js";
 export * from "./crypto.js";
 export * from "./time.js";